+++ miden-crypto/src/merkle/smt/mod.rs
@@ -121,7 +121,7 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
 
         let InnerNode { left, right } = self.get_inner_node(index.parent());
 
-        let index_is_right = index.is_value_odd();
+        let index_is_right = index.is_position_odd();
         if index_is_right { right } else { left }
     }
 
@@ -168,7 +168,7 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
     ) {
         let mut node_hash = node_hash_at_index;
         for node_depth in (0..index.depth()).rev() {
-            let is_right = index.is_value_odd();
+            let is_right = index.is_position_odd();
             index.move_up();
             let InnerNode { left, right } = self.get_inner_node(index);
             let (left, right) = if is_right {
@@ -261,7 +261,7 @@ pub(crate) trait SparseMerkleTree<const DEPTH: u8> {
 
             for node_depth in (0..node_index.depth()).rev() {
                 // Whether the node we're replacing is the right child or the left child.
-                let is_right = node_index.is_value_odd();
+                let is_right = node_index.is_position_odd();
                 node_index.move_up();
 
                 let old_node = node_mutations
@@ -553,9 +553,9 @@ impl<const DEPTH: u8> LeafIndex<DEPTH> {
         Ok(LeafIndex { index: NodeIndex::new(DEPTH, value)? })
     }
 
-    /// Returns the numeric value of this leaf index.
-    pub fn value(&self) -> u64 {
-        self.index.value()
+    /// Returns the position of this leaf index within its depth layer.
+    pub fn position(&self) -> u64 {
+        self.index.position()
     }
 }
 
@@ -585,7 +585,7 @@ impl<const DEPTH: u8> TryFrom<NodeIndex> for LeafIndex<DEPTH> {
             });
         }
 
-        Self::new(node_index.value())
+        Self::new(node_index.position())
     }
 }
 
@@ -603,7 +603,7 @@ impl<const DEPTH: u8> Deserializable for LeafIndex<DEPTH> {
 
 impl<const DEPTH: u8> Display for LeafIndex<DEPTH> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "DEPTH={}, value={}", DEPTH, self.value())
+        write!(f, "DEPTH={}, position={}", DEPTH, self.position())
     }
 }